writeGangstaModel {gangsta}	R Documentation

gangsta Utils

Description

writeGangstaModel instantiates a gangsta-derived model into computer code.

Usage


writeGangstaModel(gangstaObjects, file = file.choose())

Arguments

gangstaObjects
A list of objects of class gangsta created with functions compoundFactory and processFactory.

file
The default argument for file is file.choose(); using the default will enable the end user to overwrite an existing file or create a new file if one does not already exist. However, the end user can also specify a file on their local drive. In either case, files generated using writeGangstaModel should be saved with an ".lp" extension for compatability with lpSolve.

Value

writeGangstaModel returns a file with the simulation model code, formatted for use in lpSolve (http://lpsolve.sourceforge.net/5.5/).

[Package gangsta version 1.0 Index]